+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [Unreleased]
+- Show entities whose record has been deleted instead of raising in the UI
+
 ## [0.3.4] - 2026-03-04
 - Support an array of default objects